+++ ta4j-core/src/main/java/org/ta4j/core/criteria/ReturnOverMaxDrawdownCriterion.java
@@ -46,24 +46,28 @@ public class ReturnOverMaxDrawdownCriterion extends AbstractAnalysisCriterion {
 
     @Override
     public Num calculate(BarSeries series, Position position) {
-        var maxDrawdown = maxDrawdownCriterion.calculate(series, position);
-        if (maxDrawdown.isZero()) {
-            return NaN.NaN;
-        } else {
-            var totalProfit = netReturnCriterion.calculate(series, position);
-            return totalProfit.dividedBy(maxDrawdown);
+        if (position.isOpened()) {
+            return series.numFactory().zero();
         }
+        var maxDrawdown = maxDrawdownCriterion.calculate(series, position);
+        var netReturn = netReturnCriterion.calculate(series, position);
+        if (maxDrawdown.isZero()) {
+            return netReturn;
+        }
+        return netReturn.dividedBy(maxDrawdown);
     }
 
     @Override
     public Num calculate(BarSeries series, TradingRecord tradingRecord) {
-        var maxDrawdown = maxDrawdownCriterion.calculate(series, tradingRecord);
-        if (maxDrawdown.isZero()) {
-            return NaN.NaN;
-        } else {
-            var totalProfit = netReturnCriterion.calculate(series, tradingRecord);
-            return totalProfit.dividedBy(maxDrawdown);
+        if (tradingRecord.getPositions().isEmpty()) {
+            return series.numFactory().zero(); // penalise no-trade strategies
         }
+        var maxDrawdown = maxDrawdownCriterion.calculate(series, tradingRecord);
+        var netReturn = netReturnCriterion.calculate(series, tradingRecord);
+        if (maxDrawdown.isZero()) {
+            return netReturn; // perfect equity curve
+        }
+        return netReturn.dividedBy(maxDrawdown); // regular RoMaD
     }
 
     /** The higher the criterion value, the better. */